+++ pochidetection/cli/commands/infer.py
@@ -12,11 +12,13 @@ from pochidetection.cli.registry import resolve_infer
 from pochidetection.configs.schemas import DetectionConfigDict
 from pochidetection.interfaces.frame_sink import IFrameSink
 from pochidetection.logging import LoggerManager
-from pochidetection.scripts.common.inference import (
+from pochidetection.pipelines.builder import (
     ResolvedPipeline,
     resolve_and_setup_pipeline,
 )
-from pochidetection.scripts.common.video import (
+from pochidetection.utils import ConfigLoader
+from pochidetection.utils.config_resolver import resolve_config_path
+from pochidetection.utils.video import (
     CompositeSink,
     DisplaySink,
     FrameProcessingResult,
@@ -26,8 +28,6 @@ from pochidetection.scripts.common.video import (
     VideoWriter,
     process_frames,
 )
-from pochidetection.utils import ConfigLoader
-from pochidetection.utils.config_resolver import resolve_config_path
 
 VIDEO_EXTENSIONS = {".mp4", ".avi", ".mov"}