+++ pochidetection/pipelines/ssd_pipeline.py
@@ -8,9 +8,9 @@ from PIL import Image
 from torchvision.transforms import v2
 
 from pochidetection.core.detection import Detection
+from pochidetection.core.preprocess import gpu_preprocess_tensor
 from pochidetection.interfaces import IInferenceBackend
 from pochidetection.interfaces.pipeline import IDetectionPipeline, ImageInput
-from pochidetection.scripts.common.preprocess import gpu_preprocess_tensor
 from pochidetection.utils import PhasedTimer
 from pochidetection.utils.device import is_fp16_available